+++ modal-dialog/src/main/java/raven/modal/event/ToastCallback.java
@@ -1,0 +1,13 @@
+package raven.modal.event;
+
+/**
+ * @author Raven
+ */
+public interface ToastCallback {
+
+    void close();
+
+    void closeAllImmediately();
+
+    boolean isCurrenPromise();
+}